+++ inst/include/gdalwarpmem/gdalwarpmem.h
@@ -31,24 +31,36 @@ inline List gdal_warp_in_memory(CharacterVector source_filename,
   
   GDALDatasetH *poSrcDS;
   poSrcDS = static_cast<GDALDatasetH *>(CPLMalloc(sizeof(GDALDatasetH) * source_filename.size()));
+  bool augment;  
+  augment = !source_WKT[0].empty() || source_extent.size() == 4 || bands[0] > 0;
+  
   
   for (int i = 0; i < source_filename.size(); i++) {
-    if (!source_WKT[0].empty() || source_extent.size() == 4 || bands[0] > 0) {
+    if (augment) {
       // not dealing with subdatasets here atm
       poSrcDS[i] = gdalraster::gdalH_open_avrt(source_filename[i],   source_extent, source_WKT, 0, bands);
     } else {
       poSrcDS[i] = gdalraster::gdalH_open_dsn(source_filename[i],   0); 
+      
     }
     // unwind everything, and stop
     if (poSrcDS[i] == nullptr) {
       if (i > 0) {
         for (int j = 0; j < i; j++) GDALClose((GDALDataset *)poSrcDS[j]);
       }
-      Rprintf("input source not readable: %s\n", (char *)source_filename[i]);
+      if (augment) {
+        Rprintf("input source not readable with chosen options options: %s\n", (char *)source_filename[i]);
+        if (bands.size()>1 || bands[0] > 0) {
+          Rprintf("perhaps specified band number is not available?\n"); 
+        }
+      } else {
+       Rprintf("input source not readable: %s\n", (char *)source_filename[i]); 
+      }
+      CPLFree(poSrcDS);
       Rcpp::stop(""); 
     }
   }
-  
+
   // handle warp settings and options
   char** papszArg = nullptr;