+++ tests/testthat/test-warper.R
@@ -1,4 +1,4 @@
-f <- system.file("extdata", "sst.tif", package = "vapour")
+f <- system.file("extdata", "sst.tif", package = "vapour", mustWork = TRUE)
 
 #raster::raster(f)
 # class      : RasterLayer
@@ -92,9 +92,9 @@ test_that("robust to bad inputs", {
   expect_named(vapour_warp_raster(f, extent = c(0, 1, 0, 1),
                                   projection = "+proj=laea",  dimension = c(10, 10)))
 
-  expect_error(vapour_warp_raster(f, extent = c(0, 1, 0, 1), band = 2,
-                                  projection = "+proj=laea", source_wkt = "+proj=longlat", dimension = c(10, 10)),
-               "cannot read band 2"
+  expect_output(expect_error(vapour_warp_raster(f, extent = c(0, 1, 0, 1), band = 2,
+                                  projection = "+proj=laea", source_wkt = "+proj=longlat", dimension = c(10, 10))),
+               "band number is not available"
   )
 
 })